requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
alembic
requests
